+++ crates/ffi/src/types/image_format.rs
@@ -1,0 +1,6 @@
+#[repr(i32)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum AdImageFormat {
+    Png = 0,
+    Jpg = 1,
+}
