+++ crates/ffi/src/types/mouse_button.rs
@@ -1,0 +1,7 @@
+#[repr(i32)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum AdMouseButton {
+    Left = 0,
+    Right = 1,
+    Middle = 2,
+}
