+++ crates/ffi/src/types/mouse_event.rs
@@ -1,0 +1,11 @@
+use crate::types::mouse_button::AdMouseButton;
+use crate::types::mouse_event_kind::AdMouseEventKind;
+use crate::types::point::AdPoint;
+
+#[repr(C)]
+pub struct AdMouseEvent {
+    pub kind: AdMouseEventKind,
+    pub point: AdPoint,
+    pub button: AdMouseButton,
+    pub click_count: u32,
+}
