+++ crates/ffi/src/types/native_handle.rs
@@ -1,0 +1,4 @@
+#[repr(C)]
+pub struct AdNativeHandle {
+    pub ptr: *const std::ffi::c_void,
+}
