+++ crates/ffi/src/types/node_tree.rs
@@ -1,0 +1,7 @@
+use crate::types::node::AdNode;
+
+#[repr(C)]
+pub struct AdNodeTree {
+    pub nodes: *mut AdNode,
+    pub count: u32,
+}
