+++ crates/ffi/src/types/point.rs
@@ -1,0 +1,6 @@
+#[repr(C)]
+#[derive(Debug, Clone, Copy)]
+pub struct AdPoint {
+    pub x: f64,
+    pub y: f64,
+}
