+++ crates/ffi/src/types/rect.rs
@@ -1,0 +1,8 @@
+#[repr(C)]
+#[derive(Debug, Clone, Copy)]
+pub struct AdRect {
+    pub x: f64,
+    pub y: f64,
+    pub width: f64,
+    pub height: f64,
+}
