+++ crates/ffi/src/types/screenshot_kind.rs
@@ -1,0 +1,7 @@
+#[repr(i32)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum AdScreenshotKind {
+    Screen = 0,
+    Window = 1,
+    FullScreen = 2,
+}
