+++ crates/ffi/src/types/screenshot_target.rs
@@ -1,0 +1,8 @@
+use crate::types::screenshot_kind::AdScreenshotKind;
+
+#[repr(C)]
+pub struct AdScreenshotTarget {
+    pub kind: AdScreenshotKind,
+    pub screen_index: u64,
+    pub pid: i32,
+}
