+++ crates/ffi/src/types/scroll_params.rs
@@ -1,0 +1,7 @@
+use crate::types::direction::AdDirection;
+
+#[repr(C)]
+pub struct AdScrollParams {
+    pub direction: AdDirection,
+    pub amount: u32,
+}
