+++ crates/ffi/src/types/tree_options.rs
@@ -1,0 +1,7 @@
+#[repr(C)]
+pub struct AdTreeOptions {
+    pub max_depth: u8,
+    pub include_bounds: bool,
+    pub interactive_only: bool,
+    pub compact: bool,
+}
