+++ crates/ffi/src/types/window_info.rs
@@ -1,0 +1,13 @@
+use crate::types::rect::AdRect;
+use std::os::raw::c_char;
+
+#[repr(C)]
+pub struct AdWindowInfo {
+    pub id: *const c_char,
+    pub title: *const c_char,
+    pub app_name: *const c_char,
+    pub pid: i32,
+    pub bounds: AdRect,
+    pub has_bounds: bool,
+    pub is_focused: bool,
+}
